[refactor] Move ShardTries snapshot related code to a separate file
